Back token auth

+++ controller/user.controller.js
@@ -22,42 +22,46 @@ const editUser = (req, res, next) => {
     try {
         console.log(req.user)
         const { id } = req.params;
-        if (!isValidObjectId(id)) {
-            throw new Error('Invalid id');
+        if (id !== req.user._id) {
+            throw new Error('Cannot edit another user')
+        } else {
+            if (!isValidObjectId(id)) {
+                throw new Error('Invalid id');
+            }
+
+            const {
+                email,
+                username,
+                password,
+                avatar,
+                credit,
+                oldtrips,
+                role,
+                rating,
+                carModel,
+                carImg
+            } = req.body;
+
+            UserModel.findByIdAndUpdate(id, {
+                email,
+                username,
+                password,
+                avatar,
+                credit,
+                oldtrips,
+                role,
+                rating,
+                carModel,
+                carImg
+            }, { new: true })
+                .then(user => {
+                    if (!user) {
+                        throw new Error('User not found');
+                    }
+                    res.status(200).json(user);
+                })
+                .catch(next);
         }
-
-        const {
-            email,
-            username,
-            password,
-            avatar,
-            credit,
-            oldtrips,
-            role,
-            rating,
-            carModel,
-            carImg
-        } = req.body;
-
-        UserModel.findByIdAndUpdate(id, {
-            email,
-            username,
-            password,
-            avatar,
-            credit,
-            oldtrips,
-            role,
-            rating,
-            carModel,
-            carImg
-        }, { new: true })
-            .then(user => {
-                if (!user) {
-                    throw new Error('User not found');
-                }
-                res.status(200).json(user);
-            })
-            .catch(next);
     } catch (error) {
         res.status(400).json({ errorMessage: err.message });
     }
@@ -66,24 +70,27 @@ const editUser = (req, res, next) => {
 const deleteUser = (req, res, next) => {
     try {
         const { id } = req.params;
-        console.log(req.user)
-        if (!isValidObjectId(id)) {
-            throw new Error('Invalid id');
+        if (id !== req.user._id) {
+            throw new Error('Cannot delete another user')
+        } else {
+            if (!isValidObjectId(id)) {
+                throw new Error('Invalid id');
+            }
+            UserModel.findByIdAndDelete(id)
+                .then(user => {
+                    if (!user) {
+                        throw new Error('User not found');
+                    }
+                    return TripModel.findOneAndUpdate({ driver: { $in: [id] } }, { $pull: { driver: id } })
+                    // sacamos el driver del trip, al pintar si length 0: 'Usuario no encontrado'
+                })
+                .then(() => {
+                    return TripModel.findOneAndUpdate({ client: { $in: [id] } }, { $pull: { client: id } })
+                    // sacamos el client del trip, al pintar si length 0: 'Usuario no encontrado'
+                })
+                .then(() => res.sendStatus(204))
+                .catch(next);
         }
-        UserModel.findByIdAndDelete(id)
-            .then(user => {
-                if (!user) {
-                    throw new Error('User not found');
-                }
-                // return TripModel.findOneAndUpdate({ driver: { $in: [id] } }, { $pull: { driver: id } })
-                // sacamos el driver del trip, al pintar si no tiene nada puesto: 'Usuario no encontrado'
-            })
-            .then(() => {
-                // return TripModel.findOneAndUpdate({ client: { $in: [id] } }, { $pull: { client: id } })
-                // sacamos el client del trip, al pintar si no tiene nada puesto: 'Usuario no encontrado'
-            })
-            .then(() => res.sendStatus(204))
-            .catch(next);
     } catch (error) {
         res.status(400).json({ errorMessage: err.message });
     }
